requirements.txt
flask==2.0.1
tensorflow==2.13.0
numpy==1.24.3
opencv-python==4.8.0.74
pillow==9.5.0
gunicorn==20.1.0